feat: add command line argument to select the python module to be run

+++ caravel/run.sh
@@ -1,11 +1,18 @@
 #!/usr/bin/env bash
 set -ex
 
-mpy-cross caravel_spi.py
-mpy-cross modules/board.py
-mpy-cross modules/myspi.py
-cp ../output_files/bitstream.bin .
+module=$1
 
-sync
-mpremote mount . exec "import caravel_spi; caravel_spi.run()"
-rm bitstream.bin
+if [ -z "$module" ]; then
+    echo "Please specify the module containing run()"
+else
+
+    mpy-cross $module.py
+    mpy-cross modules/board.py
+    mpy-cross modules/myspi.py
+    cp ../output_files/bitstream.bin .
+
+    sync
+    mpremote mount . exec "import $module; $module.run()"
+    rm bitstream.bin
+fi
